Allow bad encoding when decoding bytes
`get_string` is used to forward the output streams (stdout/stderr) of called programs. If the program prints bytes that cannot be decoded, lab crashes with a `UnicodeDecodeError`.

This patch handles the issue by using an error handler while decoding (see https://docs.python.org/3/library/codecs.html#error-handlers).

A cleaner way would probably be to not decode the bytes at all an just write them to the stream as bytes. This way the streams exactly mirror the output of the program instead of modifying it to make it decodable (or crashing). But then the various streams that are used (in `call.py`, the generated `run.py`, and possibly by users of `Call` directly) would have to open files with the flag `wb` instead of `w`.